feat: implement CLI tool with configuration and file processing [deps-change]
- Add entry point for CLI tool
- Implement command-line options and argument parsing
- Create configuration validation and display commands
- Add file conversion functionality for .mmd and .mermaid files
- Introduce tests for CLI functionality and configuration handling
