+++ docs/reports/dependencies/layers.md
@@ -9,29 +9,35 @@ flowchart LR
     classDef dir fill:#0000,stroke:#999
     subgraph src["src"]
         src/domain["/domain"]:::dir
-        src/application["/application"]:::dir
         src/infrastructure["/infrastructure"]:::dir
+        src/application["/application"]:::dir
         src/presentation["/presentation"]:::dir
+        src/index.ts["index.ts"]
+        subgraph src/types["/types"]
+            src/types/config.ts["config.ts"]
+        end
     end
     subgraph node//modules["node_modules"]
         node//modules/valibot/dist/index.d.cts["valibot"]
-        node//modules/globby/index.d.ts["globby"]
         node//modules/cosmiconfig/dist/index.d.ts["cosmiconfig"]
         node//modules/cosmiconfig//typescript//loader/dist/types/index.d.ts["cosmiconfig-typescript-loader"]
+        node//modules/globby/index.d.ts["globby"]
         node//modules/commander/typings/index.d.ts["commander"]
     end
     src/domain-->node//modules/valibot/dist/index.d.cts
-    src/application-->src/domain
-    src/infrastructure-->node//modules/globby/index.d.ts
-    src/application-->src/infrastructure
     src/infrastructure-->node//modules/cosmiconfig/dist/index.d.ts
     src/infrastructure-->node//modules/cosmiconfig//typescript//loader/dist/types/index.d.ts
     src/infrastructure-->src/domain
     src/infrastructure-->package.json
+    src/infrastructure-->node//modules/globby/index.d.ts
+    src/application-->src/domain
+    src/application-->src/infrastructure
     src/presentation-->src/application
     src/presentation-->src/domain
-    src/presentation-->src/infrastructure
     src/presentation-->node//modules/commander/typings/index.d.ts
     src/presentation-->node//modules/valibot/dist/index.d.cts
+    src/presentation-->src/infrastructure
+    src/index.ts-->src/presentation
+    src/types/config.ts-->src/domain
 ```
