+++ src/index.ts
@@ -1,0 +1,9 @@
+#!/usr/bin/env bun
+
+/**
+ * Entry point for the mermaid-markdown-wrap CLI tool
+ * This file serves as the single entry point that delegates to the CLI implementation
+ */
+
+// Import and run the CLI
+import "./presentation/cli.js";
